+++ test-browser/isolated-tests/services/field-mapper.spec.ts
@@ -48,16 +48,16 @@ describe('ElementMapper', () => {
 
         elementMapper.mapElements({
             item: item,
-            modularContent: {},
             preparedItems: {
                 'cd': {
                     _raw: null as any,
-                    system: systemData as any
+                    system: systemData as any,
+                    getAllElements: () => []
                 }
             },
             processedItems: {},
             processingStartedForCodenames: [],
-            queryConfig: {}
+            queryConfig: {},
         });
 
         expect(console.warn).toHaveBeenCalledTimes(1);