+++ test-browser/live-tests/retry/retry-404-live.spec.ts
@@ -3,12 +3,16 @@ import { retryService } from '@kentico/kontent-core';
 import { Context, setup } from '../../setup';
 
 describe('Live 404 retry', () => {
-    const context = new Context();
-
     const retryAttempts = 3;
 
-    // set retry attempts
-    context.retryAttempts = retryAttempts;
+    const context = new Context();
+    context.retryStrategy = {
+        maxAttempts: retryAttempts,
+        addJitter: false,
+        deltaBackoffMs: 1000,
+        maxCumulativeWaitTimeMs: 50000,
+        canRetryError: (error) => true
+    };
 
     // set fake base url because we want this to fail
     context.baseUrl = 'http://fakeurl';
@@ -30,8 +34,8 @@ describe('Live 404 retry', () => {
             });
     });
 
-    it(`Warning for retry attempt should have been called '0' times because 404 response codes should not be retried`, () => {
-        expect(retryService.debugLogAttempt).toHaveBeenCalledTimes(0);
+    it(`Warning for retry attempt should have been called '${retryAttempts}' times`, () => {
+        expect(retryService.debugLogAttempt).toHaveBeenCalledTimes(retryAttempts);
     });
 });