+++ test-browser/live-tests/retry/retry-isolated-rxjs-do-retry.spec.ts
@@ -2,43 +2,43 @@ import { retryService } from '@kentico/kontent-core';
 
 import { Context, setup } from '../../setup';
 
-describe('Retry - isolated - do not retry', () => {
-    const retryAttempts = 0;
+describe('Retry rxjs - isolated - retry', () => {
+    const retryAttempts = 2;
     const MAX_SAFE_TIMEOUT = Math.pow(2, 31) - 1;
 
     jasmine.DEFAULT_TIMEOUT_INTERVAL = MAX_SAFE_TIMEOUT;
 
-    beforeAll(done => {
+    beforeAll((done) => {
         spyOn(retryService, 'debugLogAttempt').and.callThrough();
 
         const context = new Context();
         context.retryStrategy = {
-            maxAttempts: 3,
+            maxAttempts: retryAttempts,
             addJitter: false,
             deltaBackoffMs: 1000,
             maxCumulativeWaitTimeMs: 5000,
-            canRetryError: (error) => false
+            canRetryError: (error) => true
         };
 
         setup(context);
         const client = context.deliveryClient;
 
-        const observable = client
-            .items()
-            .withUrl('fakeUrl')
-            .toObservable();
+        const observable = client.item('xxxyyy').toObservable(); // throws 404 which we retry
 
         observable.subscribe(
             response => {
-                done();
+                throw Error(`This call should not succeed`);
             },
             error => {
+                console.warn(error);
                 done();
             }
         );
+
     });
 
-    it(`Warning for retry attempt should have been called '${retryAttempts}' times`, () => {
-        expect(retryService.debugLogAttempt).toHaveBeenCalledTimes(0);
+    it(`Warning for retry attempt should have been called '${retryAttempts}'`, () => {
+        expect(retryService.debugLogAttempt).toHaveBeenCalledTimes(retryAttempts);
     });
 });
+